refactor: moved promotion component inside cart components as it only used there.
